+++ pdfs/test-envelope-123456.pdf
@@ -1,0 +1,1 @@
+Test PDF content
